# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: API
  version: 1.0.0
  description: >
    API documentation for API services. The shareholding endpoint below
    now surfaces ownership data sourced from Yahoo Finance via RapidAPI.
servers:
  - url: /api
    description: Application API
paths:
  /shareholdings:
    get:
      summary: Retrieve shareholding information for a company
      operationId: getShareholding
      tags:
        - Shareholding
      parameters:
        - name: symbol
          in: query
          description: Stock ticker symbol as recognised by Yahoo Finance (e.g. AAPL).
          required: true
          schema:
            type: string
        - name: limit
          in: query
          description: Optional maximum number of holders to keep per ownership list (1-100).
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: refresh
          in: query
          description: Set to true to fetch fresh data from Yahoo Finance and store a new snapshot.
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: Shareholding information retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareholdingResponse'
              examples:
                sample:
                  value:
                    summaryDetail:
                      maxAge: 1
                      priceHint: 2
                      previousClose: 202.8099
                      open: 206.45
                      dayLow: 202.07
                      dayHigh: 207.97
                      regularMarketPreviousClose: 202.8099
                      regularMarketOpen: 206.45
                      regularMarketDayLow: 202.07
                      regularMarketDayHigh: 207.97
                      dividendRate: 0.04
                      dividendYield: 0.0002
                      exDividendDate: 1757548800
                      payoutRatio: 0.0114
                      fiveYearAvgDividendYield: 0.06
                      beta: 2.123
                      trailingPE: 57.854286
                      forwardPE: 49.14806
                      volume: 179802178
                      regularMarketVolume: 179802178
                      averageVolume: 176961839
                      averageVolume10days: 177631280
                      averageDailyVolume10Day: 177631280
                      bid: 202.21
                      ask: 203.29
                      bidSize: 100
                      askSize: 100
                      marketCap: 4930024243200
                      fiftyTwoWeekLow: 86.62
                      fiftyTwoWeekHigh: 212.19
                      allTimeHigh: 212.19
                      allTimeLow: 0.033333
                      priceToSalesTrailing12Months: 29.839512
                      fiftyDayAverage: 181.9192
                      twoHundredDayAverage: 148.2147
                      trailingAnnualDividendRate: 0.04
                      trailingAnnualDividendYield: 0.00019722902
                      currency: USD
                      fromCurrency: null
                      toCurrency: null
                      lastMarket: null
                      coinMarketCapLink: null
                      algorithm: null
                      tradeable: false
                    institutionOwnership:
                      maxAge: 1
                      ownershipList:
                        - maxAge: 1
                          reportDate:
                            raw: 1751241600
                            fmt: "2025-06-30"
                          organization: "Vanguard Group Inc"
                          pctHeld:
                            raw: 0.0919
                            fmt: "9.19%"
                          position:
                            raw: 2232444958
                            fmt: "2.23B"
                            longFmt: "2,232,444,958"
                          value:
                            raw: 452047791808
                            fmt: "452.05B"
                            longFmt: "452,047,791,808"
                          pctChange:
                            raw: 0.018
                            fmt: "1.80%"
                        - maxAge: 1
                          reportDate:
                            raw: 1751241600
                            fmt: "2025-06-30"
                          organization: "Blackrock Inc."
                          pctHeld:
                            raw: 0.078600004
                            fmt: "7.86%"
                          position:
                            raw: 1909788972
                            fmt: "1.91B"
                            longFmt: "1,909,788,972"
                          value:
                            raw: 386713179431
                            fmt: "386.71B"
                            longFmt: "386,713,179,431"
                          pctChange:
                            raw: 0.0042
                            fmt: "0.42%"
                        - maxAge: 1
                          reportDate:
                            raw: 1751241600
                            fmt: "2025-06-30"
                          organization: "FMR, LLC"
                          pctHeld:
                            raw: 0.041100003
                            fmt: "4.11%"
                          position:
                            raw: 997977839
                            fmt: "997.98M"
                            longFmt: "997,977,839"
                          value:
                            raw: 202080538101
                            fmt: "202.08B"
                            longFmt: "202,080,538,101"
                          pctChange:
                            raw: -0.0039
                            fmt: "-0.39%"
                        - maxAge: 1
                          reportDate:
                            raw: 1751241600
                            fmt: "2025-06-30"
                          organization: "State Street Corporation"
                          pctHeld:
                            raw: 0.0403
                            fmt: "4.03%"
                          position:
                            raw: 978208862
                            fmt: "978.21M"
                            longFmt: "978,208,862"
                          value:
                            raw: 198077517839
                            fmt: "198.08B"
                            longFmt: "198,077,517,839"
                          pctChange:
                            raw: 0.0099
                            fmt: "0.99%"
                        - maxAge: 1
                          reportDate:
                            raw: 1751241600
                            fmt: "2025-06-30"
                          organization: "Geode Capital Management, LLC"
                          pctHeld:
                            raw: 0.0238
                            fmt: "2.38%"
                          position:
                            raw: 579213497
                            fmt: "579.21M"
                            longFmt: "579,213,497"
                          value:
                            raw: 117284944189
                            fmt: "117.28B"
                            longFmt: "117,284,944,189"
                          pctChange:
                            raw: 0.0149
                            fmt: "1.49%"
                        - maxAge: 1
                          reportDate:
                            raw: 1751241600
                            fmt: "2025-06-30"
                          organization: "JPMORGAN CHASE & CO"
                          pctHeld:
                            raw: 0.019
                            fmt: "1.90%"
                          position:
                            raw: 462586867
                            fmt: "462.59M"
                            longFmt: "462,586,867"
                          value:
                            raw: 93669217239
                            fmt: "93.67B"
                            longFmt: "93,669,217,239"
                          pctChange:
                            raw: 0.098500006
                            fmt: "9.85%"
                        - maxAge: 1
                          reportDate:
                            raw: 1751241600
                            fmt: "2025-06-30"
                          organization: "Price (T.Rowe) Associates Inc"
                          pctHeld:
                            raw: 0.0167
                            fmt: "1.67%"
                          position:
                            raw: 406325355
                            fmt: "406.33M"
                            longFmt: "406,325,355"
                          value:
                            raw: 82276823365
                            fmt: "82.28B"
                            longFmt: "82,276,823,365"
                          pctChange:
                            raw: 0.0836
                            fmt: "8.36%"
                        - maxAge: 1
                          reportDate:
                            raw: 1751241600
                            fmt: "2025-06-30"
                          organization: "NORGES BANK"
                          pctHeld:
                            raw: 0.0134000005
                            fmt: "1.34%"
                          position:
                            raw: 325253897
                            fmt: "325.25M"
                            longFmt: "325,253,897"
                          value:
                            raw: 65860663390
                            fmt: "65.86B"
                            longFmt: "65,860,663,390"
                          pctChange:
                            raw: 0.0037
                            fmt: "0.37%"
                        - maxAge: 1
                          reportDate:
                            raw: 1751241600
                            fmt: "2025-06-30"
                          organization: "Morgan Stanley"
                          pctHeld:
                            raw: 0.0134000005
                            fmt: "1.34%"
                          position:
                            raw: 325710808
                            fmt: "325.71M"
                            longFmt: "325,710,808"
                          value:
                            raw: 65953183301
                            fmt: "65.95B"
                            longFmt: "65,953,183,301"
                          pctChange:
                            raw: -0.0537
                            fmt: "-5.37%"
                        - maxAge: 1
                          reportDate:
                            raw: 1751241600
                            fmt: "2025-06-30"
                          organization: "Northern Trust Corporation"
                          pctHeld:
                            raw: 0.0107
                            fmt: "1.07%"
                          position:
                            raw: 261151459
                            fmt: "261.15M"
                            longFmt: "261,151,459"
                          value:
                            raw: 52880560367
                            fmt: "52.88B"
                            longFmt: "52,880,560,367"
                          pctChange:
                            raw: -0.0070999996
                            fmt: "-0.71%"
                    majorHoldersBreakdown:
                      maxAge: 1
                      insidersPercentHeld: 0.04321
                      institutionsPercentHeld: 0.69014
                      institutionsFloatPercentHeld: 0.72130996
                      institutionsCount: 6801
                    price:
                      maxAge: 1
                      preMarketSource: "FREE_REALTIME"
                      postMarketChangePercent: 0.00064195576
                      postMarketChange: 0.12998962
                      postMarketTime: 1761955199
                      postMarketPrice: 202.62
                      postMarketSource: "FREE_REALTIME"
                      regularMarketChangePercent: -0.0015773417
                      regularMarketChange: -0.3199005
                      regularMarketTime: 1761940800
                      priceHint: 2
                      regularMarketPrice: 202.49
                      regularMarketDayHigh: 207.97
                      regularMarketDayLow: 202.07
                      regularMarketVolume: 179802178
                      averageDailyVolume10Day: 177631280
                      averageDailyVolume3Month: 176961839
                      regularMarketPreviousClose: 202.8099
                      regularMarketSource: "FREE_REALTIME"
                      regularMarketOpen: 206.45
                      exchange: "NMS"
                      exchangeName: "NasdaqGS"
                      exchangeDataDelayedBy: 0
                      marketState: "CLOSED"
                      quoteType: "EQUITY"
                      symbol: "NVDA"
                      underlyingSymbol: null
                      shortName: "NVIDIA Corporation"
                      longName: "NVIDIA Corporation"
                      currency: "USD"
                      quoteSourceName: "Nasdaq Real Time Price"
                      currencySymbol: "$"
                      fromCurrency: null
                      toCurrency: null
                      lastMarket: null
                      marketCap: 4930024243200
                    majorDirectHolders:
                      holders: []
                      maxAge: 1
                    netSharePurchaseActivity:
                      maxAge: 1
                      period: "6m"
                      buyInfoCount: 18
                      buyInfoShares: 4063460
                      buyPercentInsiderShares: 0.004
                      sellInfoCount: 54
                      sellInfoShares: 11500037
                      sellPercentInsiderShares: 0.011
                      netInfoCount: 72
                      netInfoShares: -7436577
                      netPercentInsiderShares: -0.0069999998
                      totalInsiderShares: 1050002944
                    quoteType:
                      exchange: "NMS"
                      quoteType: "EQUITY"
                      symbol: "NVDA"
                      underlyingSymbol: "NVDA"
                      shortName: "NVIDIA Corporation"
                      longName: "NVIDIA Corporation"
                      firstTradeDateEpochUtc: 917015400
                      timeZoneFullName: "America/New_York"
                      timeZoneShortName: "EST"
                      uuid: "7f5f6a07-b148-30f4-98a2-2caa3df2aed0"
                      messageBoardId: "finmb_32307"
                      gmtOffSetMilliseconds: -18000000
                      maxAge: 1
                    fundOwnership:
                      maxAge: 1
                      ownershipList:
                        - maxAge: 1
                          reportDate:
                            raw: 1751241600
                            fmt: "2025-06-30"
                          organization: "VANGUARD INDEX FUNDS-Vanguard Total Stock Market Index Fund"
                          pctHeld:
                            raw: 0.0307
                            fmt: "3.07%"
                          position:
                            raw: 745382412
                            fmt: "745.38M"
                            longFmt: "745,382,412"
                          value:
                            raw: 150932488700
                            fmt: "150.93B"
                            longFmt: "150,932,488,700"
                          pctChange:
                            raw: 0.0178
                            fmt: "1.78%"
                        - maxAge: 1
                          reportDate:
                            raw: 1751241600
                            fmt: "2025-06-30"
                          organization: "VANGUARD INDEX FUNDS-Vanguard 500 Index Fund"
                          pctHeld:
                            raw: 0.028499998
                            fmt: "2.85%"
                          position:
                            raw: 692239585
                            fmt: "692.24M"
                            longFmt: "692,239,585"
                          value:
                            raw: 140171597369
                            fmt: "140.17B"
                            longFmt: "140,171,597,369"
                          pctChange:
                            raw: 0.0173
                            fmt: "1.73%"
                        - maxAge: 1
                          reportDate:
                            raw: 1756598400
                            fmt: "2025-08-31"
                          organization: "Fidelity Concord Street Trust-Fidelity 500 Index Fund"
                          pctHeld:
                            raw: 0.0127
                            fmt: "1.27%"
                          position:
                            raw: 309653955
                            fmt: "309.65M"
                            longFmt: "309,653,955"
                          value:
                            raw: 62701831048
                            fmt: "62.7B"
                            longFmt: "62,701,831,048"
                          pctChange:
                            raw: 0.0017
                            fmt: "0.17%"
                        - maxAge: 1
                          reportDate:
                            raw: 1759190400
                            fmt: "2025-09-30"
                          organization: "iShares Trust-iShares Core S&P 500 ETF"
                          pctHeld:
                            raw: 0.0123000005
                            fmt: "1.23%"
                          position:
                            raw: 299278139
                            fmt: "299.28M"
                            longFmt: "299,278,139"
                          value:
                            raw: 60600832010
                            fmt: "60.6B"
                            longFmt: "60,600,832,010"
                          pctChange:
                            raw: 0.0198
                            fmt: "1.98%"
                        - maxAge: 1
                          reportDate:
                            raw: 1759190400
                            fmt: "2025-09-30"
                          organization: "SPDR S&P 500 ETF TRUST"
                          pctHeld:
                            raw: 0.012
                            fmt: "1.20%"
                          position:
                            raw: 290940470
                            fmt: "290.94M"
                            longFmt: "290,940,470"
                          value:
                            raw: 58912537368
                            fmt: "58.91B"
                            longFmt: "58,912,537,368"
                          pctChange:
                            raw: -0.012200001
                            fmt: "-1.22%"
                        - maxAge: 1
                          reportDate:
                            raw: 1751241600
                            fmt: "2025-06-30"
                          organization: "VANGUARD INDEX FUNDS-Vanguard Growth Index Fund"
                          pctHeld:
                            raw: 0.0095
                            fmt: "0.95%"
                          position:
                            raw: 230074826
                            fmt: "230.07M"
                            longFmt: "230,074,826"
                          value:
                            raw: 46587852780
                            fmt: "46.59B"
                            longFmt: "46,587,852,780"
                          pctChange:
                            raw: 0.039100002
                            fmt: "3.91%"
                        - maxAge: 1
                          reportDate:
                            raw: 1759190400
                            fmt: "2025-09-30"
                          organization: "Invesco QQQ Trust, Series 1"
                          pctHeld:
                            raw: 0.0084
                            fmt: "0.84%"
                          position:
                            raw: 204053116
                            fmt: "204.05M"
                            longFmt: "204,053,116"
                          value:
                            raw: 41318716579
                            fmt: "41.32B"
                            longFmt: "41,318,716,579"
                          pctChange:
                            raw: 0.0052
                            fmt: "0.52%"
                        - maxAge: 1
                          reportDate:
                            raw: 1751241600
                            fmt: "2025-06-30"
                          organization: "VANGUARD INSTITUTIONAL INDEX FUNDS-Vanguard Institutional Index Fund"
                          pctHeld:
                            raw: 0.0062
                            fmt: "0.62%"
                          position:
                            raw: 149474240
                            fmt: "149.47M"
                            longFmt: "149,474,240"
                          value:
                            raw: 30267039678
                            fmt: "30.27B"
                            longFmt: "30,267,039,678"
                          pctChange:
                            raw: -0.0099
                            fmt: "-0.99%"
                        - maxAge: 1
                          reportDate:
                            raw: 1756598400
                            fmt: "2025-08-31"
                          organization: "VANGUARD WORLD FUND-Vanguard Information Technology Index Fund"
                          pctHeld:
                            raw: 0.0047999998
                            fmt: "0.48%"
                          position:
                            raw: 116052317
                            fmt: "116.05M"
                            longFmt: "116,052,317"
                          value:
                            raw: 23499434306
                            fmt: "23.5B"
                            longFmt: "23,499,434,306"
                          pctChange:
                            raw: -0.0293
                            fmt: "-2.93%"
                        - maxAge: 1
                          reportDate:
                            raw: 1759190400
                            fmt: "2025-09-30"
                          organization: "GROWTH FUND OF AMERICA"
                          pctHeld:
                            raw: 0.0043
                            fmt: "0.43%"
                          position:
                            raw: 105286193
                            fmt: "105.29M"
                            longFmt: "105,286,193"
                          value:
                            raw: 21319401798
                            fmt: "21.32B"
                            longFmt: "21,319,401,798"
                          pctChange:
                            raw: 0.0072000003
                            fmt: "0.72%"
                    symbol: "NVDA"
        '400':
          description: Yahoo Finance rejected the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error in provided query parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '500':
          description: Unexpected error while fetching shareholding data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /health:
    get:
      summary: Health check endpoint
      operationId: getHealth
      tags:
        - Health
      responses:
        '200':
          description: Service is online
          content:
            text/plain:
              schema:
                type: string
                example: online
components:
  schemas:
    ShareholdingResponse:
      type: object
      properties:
        summaryDetail:
          $ref: '#/components/schemas/YahooSummaryDetail'
        institutionOwnership:
          $ref: '#/components/schemas/YahooOwnershipCollection'
        majorHoldersBreakdown:
          $ref: '#/components/schemas/MajorHoldersBreakdown'
        price:
          $ref: '#/components/schemas/YahooPriceDetail'
        majorDirectHolders:
          $ref: '#/components/schemas/MajorDirectHolders'
        netSharePurchaseActivity:
          $ref: '#/components/schemas/NetSharePurchaseActivity'
        quoteType:
          $ref: '#/components/schemas/YahooQuoteType'
        fundOwnership:
          $ref: '#/components/schemas/YahooOwnershipCollection'
        symbol:
          type: string
          description: Echo of the requested ticker symbol.
      additionalProperties: false
      required:
        - symbol
    YahooSummaryDetail:
      type: object
      description: Key trading statistics for the requested symbol.
      additionalProperties: true
      properties:
        maxAge:
          type: integer
        previousClose:
          type: number
          format: float
        open:
          type: number
          format: float
        dayLow:
          type: number
          format: float
        dayHigh:
          type: number
          format: float
        currency:
          type: string
        marketCap:
          type: number
          format: float
    YahooOwnershipCollection:
      type: object
      description: Ownership information grouped by holder type.
      properties:
        maxAge:
          type: integer
        ownershipList:
          type: array
          items:
            $ref: '#/components/schemas/YahooOwnershipEntry'
      required:
        - ownershipList
      additionalProperties: false
    YahooOwnershipEntry:
      type: object
      properties:
        maxAge:
          type: integer
        reportDate:
          $ref: '#/components/schemas/YahooReportDate'
        organization:
          type: string
        pctHeld:
          $ref: '#/components/schemas/YahooFormattedValue'
        position:
          $ref: '#/components/schemas/YahooFormattedValue'
        value:
          $ref: '#/components/schemas/YahooFormattedValue'
        pctChange:
          $ref: '#/components/schemas/YahooFormattedValue'
      additionalProperties: false
    YahooReportDate:
      type: object
      properties:
        raw:
          type: integer
          description: Unix timestamp of the reporting date.
        fmt:
          type: string
          format: date
          description: Formatted reporting date.
      additionalProperties: false
    YahooFormattedValue:
      type: object
      properties:
        raw:
          type: number
        fmt:
          type: string
        longFmt:
          type: string
      additionalProperties: false
    MajorHoldersBreakdown:
      type: object
      properties:
        maxAge:
          type: integer
        insidersPercentHeld:
          type: number
          format: float
        institutionsPercentHeld:
          type: number
          format: float
        institutionsFloatPercentHeld:
          type: number
          format: float
        institutionsCount:
          type: integer
      additionalProperties: false
    YahooPriceDetail:
      type: object
      description: Real-time price snapshot for the symbol.
      properties:
        maxAge:
          type: integer
        regularMarketPrice:
          type: number
          format: float
        regularMarketChange:
          type: number
          format: float
        regularMarketChangePercent:
          type: number
          format: float
        regularMarketDayHigh:
          type: number
          format: float
        regularMarketDayLow:
          type: number
          format: float
        regularMarketVolume:
          type: integer
        exchangeName:
          type: string
        marketState:
          type: string
        quoteType:
          type: string
        symbol:
          type: string
        shortName:
          type: string
        currency:
          type: string
      additionalProperties: true
    MajorDirectHolders:
      type: object
      properties:
        holders:
          type: array
          items:
            type: object
        maxAge:
          type: integer
      additionalProperties: false
    NetSharePurchaseActivity:
      type: object
      properties:
        maxAge:
          type: integer
        period:
          type: string
        buyInfoCount:
          type: integer
        sellInfoCount:
          type: integer
        netInfoCount:
          type: integer
        netInfoShares:
          type: integer
        totalInsiderShares:
          type: integer
      additionalProperties: false
    YahooQuoteType:
      type: object
      properties:
        exchange:
          type: string
        quoteType:
          type: string
        symbol:
          type: string
        underlyingSymbol:
          type: string
          nullable: true
        shortName:
          type: string
        longName:
          type: string
        timeZoneFullName:
          type: string
        timeZoneShortName:
          type: string
        gmtOffSetMilliseconds:
          type: integer
        firstTradeDateEpochUtc:
          type: integer
        messageBoardId:
          type: string
        uuid:
          type: string
      additionalProperties: true
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        error:
          description: Additional context returned by the upstream provider.
          nullable: true
      required:
        - message
    ValidationErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: The given data was invalid.
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
      required:
        - message
        - errors
tags:
  - name: Shareholding
    description: Access to Yahoo Finance ownership and price data via RapidAPI.
  - name: Health
    description: Health check endpoints.